python_service/requirements.txt
# requirements.txt
grpcio
grpcio-tools